chore(types): pyright 타입 불일치 정리
에디터(Pylance) 에러 제거 목적, 런타임 동작 불변.

- Mr2sSolverProtocol.evaluator 를 read-only property 로 선언 — mutable
  속성 선언이면 DnCMr2sSolver 의 property 구현이 구조적으로 불일치
- Edge.weight / AdjEntry.weight / get_indicator_function weight 를
  float 로 — harmonic super edge 가중치가 실수라 int 힌트는 거짓
- chain_contraction 의 Graph 생성을 dict 전달로 통일
- BinaryPolynomial 을 dimod.higherorder.polynomial 에서 직접 import

Refs #73

+++ mr2s_module/util/qubo_util.py
@@ -3,7 +3,7 @@ from dwave.samplers import SimulatedAnnealingSampler
 
 sampler = SimulatedAnnealingSampler()
 
-def get_indicator_function(i: int, j: int, edge_id: int, weight: int) -> BinaryPolynomial:
+def get_indicator_function(i: int, j: int, edge_id: int, weight: float) -> BinaryPolynomial:
   if i == j:
     raise ValueError(f"i and j must be different, but both are {i}")
